requirements.txt
# Requirements used for managing tests, builds, documentation
tox
wheel
twine

# Requirements for executing tests in test environments
#coverage
#mock>=1.0.1
pytest-django>=3.0.0
#pytest-cov>=2.4.0
flake8>=2.4.1
